Move component code and html to a different file
